Move Hashable/HashArray fixtures to the fixtures NS
