+++ tests/fixtures/HashArrayElement.php
@@ -1,7 +1,11 @@
 <?php
 
-namespace Wikibase\Test\HashArray;
+namespace Wikibase\Test\DataModel\Fixtures;
 
+/**
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
 class HashArrayElement implements \Hashable {
 
 	public $text = '';
