+++ tests/fixtures/MutableHashable.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Wikibase\Test\DataModel\Fixtures;
+
+use Hashable;
+
+/**
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
+class MutableHashable implements Hashable {
+
+	public $text = '';
+
+	public function getHash() {
+		return sha1( __CLASS__ . $this->text );
+	}
+
+}
